Remove unnecessary casts to unsigned int for uid_t.
Bug 14753

+++ src/api/allocate.c
@@ -1172,7 +1172,7 @@ _handle_msg(slurm_msg_t *msg, uint16_t msg_type, void **resp)
 	if ((req_uid != slurm_conf.slurm_user_id) && (req_uid != 0) &&
 	    (req_uid != uid)) {
 		error ("Security violation, slurm message from uid %u",
-			(unsigned int) req_uid);
+		       req_uid);
 		return 0;
 	}
